refactor: 미들웨어 함수 분리
